fix a bug in MedFeatures read from csv

+++ Internal/MedProcessTools/MedProcessTools/MedFeatures.cpp
@@ -387,13 +387,14 @@ int MedFeatures::read_from_csv_mat(const string &csv_fname)
 			newSample.outcome = stof(fields[idx++]);
 			newSample.outcomeTime = stoi(fields[idx++]);
 			newSample.split = stoi(fields[idx++]);
+
 			newSample.prediction.resize(max_pred + 1);
-			for (size_t i = 0; i <= max_pred; ++i)
+
+			for (int i = 0; i <= max_pred; ++i)
 				if (field_ind_to_pred_index[i] >= 0)
 					newSample.prediction[i] = stof(fields[field_ind_to_pred_index[i]]);
-
 			samples.push_back(newSample);
-
+			
 			for (int i = 0; i < names.size(); i++)
 				data[names[i]].push_back(stof(fields[idx + i]));
 		}
